fixup imports for python 3.4
